@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\SchoolClass;
+use App\Models\SchoolClass;
 use Illuminate\Http\Request;
 
 class SchoolClassController extends Controller
@@ -10,13 +10,14 @@ class SchoolClassController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function index()
     {
-        $classesIntranet = SchoolClass::fetchClassesFromIntranet();
-        $classes = SchoolClass::identifyClass($classesIntranet);
-        return view('classes.index', compact('classes', 'classesIntranet'));
+
+        $classes = SchoolClass::all();
+
+        return view('classes.index', compact('classes'));
 
     }
 
@@ -34,29 +35,24 @@ class SchoolClassController extends Controller
      * Store a newly created resource in storage.
      *
      * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
+     *
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(Request $request)
     {
+        $classes = SchoolClass::getClasses();
         $classesIntranet = SchoolClass::fetchClassesFromIntranet();
-        $classes = SchoolClass::identifyClass($classesIntranet);
-        $selected = $_POST;
         foreach ($classesIntranet as $class) {
-            if (array_key_exists($class['name'], $selected)) {
-                SchoolClass::removeOldClasses($classes,$selected);
-                SchoolClass::synchronise($class);
-            }
+            SchoolClass::synchronise($class);
         }
-        /*$old = SchoolClass::where('name',$class['name']);
-        $old->delete();*/
         return redirect()->route('classes.index');
-
     }
 
     /**
      * Display the specified resource.
      *
-     * @param \App\SchoolClass $schoolClass
+     * @param \App\Models\SchoolClass $schoolClass
+     *
      * @return \Illuminate\Http\Response
      */
     public function show(SchoolClass $schoolClass)
@@ -67,7 +63,8 @@ class SchoolClassController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param \App\SchoolClass $schoolClass
+     * @param \App\Models\SchoolClass $schoolClass
+     *
      * @return \Illuminate\Http\Response
      */
     public function edit(SchoolClass $schoolClass)
@@ -79,7 +76,8 @@ class SchoolClassController extends Controller
      * Update the specified resource in storage.
      *
      * @param \Illuminate\Http\Request $request
-     * @param \App\SchoolClass $schoolClass
+     * @param \App\Models\SchoolClass  $schoolClass
+     *
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $schoolClasses)
@@ -90,7 +88,8 @@ class SchoolClassController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param \App\SchoolClass $schoolClass
+     * @param \App\Models\SchoolClass $schoolClass
+     *
      * @return \Illuminate\Http\Response
      */
     public function destroy(SchoolClass $schoolClass)
