@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Role;
+use App\Models\User;
 use Illuminate\Http\Request;
-use App\User;
-use App\Role;
 use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
