@@ -2,7 +2,7 @@
 
 namespace App\Http\Middleware;
 
-use App\User;
+use App\Models\User;
 use Illuminate\Support\Facades\Auth;
 use Closure;
 use Illuminate\Support\Facades\Redirect;
@@ -23,19 +23,19 @@ class DevLogin
             $user = User::find(env('USER_ID'));
             if ($user){
                 if (Auth::attempt(['username' => $user->username, 'password' => 'Pa$$w0rd'])){
-                    \Session::put('isDev', true);  
+                    \Session::put('isDev', true);
                     return $next($request);
-                }else{  
-                    \Session::put('isDev', false);   
+                }else{
+                    \Session::put('isDev', false);
                     Redirect::guest('login');
                 }
             }else{
-                \Session::put('isDev', false);    
+                \Session::put('isDev', false);
                 Redirect::to('events');
             }
-            
-        }else{ 
-            \Session::put('isDev', false);  
+
+        }else{
+            \Session::put('isDev', false);
             error_log("Prod");
         }
 
