@@ -2,18 +2,19 @@
 
 namespace App\Providers;
 
+// use Illuminate\Support\Facades\Gate;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Gate;
 
 class AuthServiceProvider extends ServiceProvider
 {
     /**
-     * The policy mappings for the application.
+     * The model to policy mappings for the application.
      *
-     * @var array
+     * @var array<class-string, class-string>
      */
     protected $policies = [
-        // 'App\Model' => 'App\Policies\ModelPolicy',
+        // 'App\Models\Model' => 'App\Policies\ModelPolicy',
     ];
 
     /**
@@ -24,7 +25,6 @@ class AuthServiceProvider extends ServiceProvider
     public function boot()
     {
         $this->registerPolicies();
-
         Gate::define('isGest', function ($user) {
             return $user->role->slug ==  'GEST';
         });
@@ -38,7 +38,8 @@ class AuthServiceProvider extends ServiceProvider
         });
 
         Gate::define('isOrg', function ($user) {
-            return $user->role->slug ==  'ORG';
+            return ($user->role->slug ==  'ORG' || $user->role->slug == 'ADMIN');
         });
+
     }
 }
