@@ -2,12 +2,8 @@
 
 namespace Database\Seeders;
 
-use App\Court;
-use Illuminate\Database\Seeder;
-
-use Illuminate\Support\Facades\Session;
 use Database\Seeders\Natives;
-use Laravel\SerializableClosure\Serializers\Native;
+use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
 {
@@ -19,13 +15,12 @@ class DatabaseSeeder extends Seeder
     public function run()
     {
         $this->call(Natives\RoleSeeder::class);
-        $this->call(Natives\AdminUserSeeder::class);
         $this->call(Natives\GameTypeSeeder::class);
         $this->call(Natives\PoolModeSeeder::class);
         $this->call(Natives\SportSeeder::class);
         $this->call(Natives\CourtSeeder::class);
-        $this->call(Natives\SchoolClassSeeder::class);
         if (env('APP_ENV') == 'local') {
+            $this->call(Natives\AdminUserSeeder::class);
             $this->call(Fakes\Joutes2022Seeder::class);
             $this->call(Fakes\FakeResultPoolState::class);
             $this->call(Fakes\UserTableSeeder::class);
