@@ -2,11 +2,10 @@
 
 namespace Database\Seeders\Fakes;
 
-use App\Role;
-use App\Team;
-use App\GameType;
+use App\Models\GameType;
+use App\Models\Team;
+use App\Models\Tournament;
 use Carbon\Carbon;
-use App\Tournament;
 use Illuminate\Database\Seeder;
 
 class Joutes2020Seeder extends Seeder
@@ -24,10 +23,10 @@ class Joutes2020Seeder extends Seeder
         $user = \Config::get('database.connections.mysql.username');
         $pass = \Config::get('database.connections.mysql.password');
 
-        $this->event = \App\Event::where('name', 'like', '%2019%')->first();
+        $this->event = \App\Models\Event::where('name', 'like', '%2019%')->first();
         if ($this->event) die("L'événement existe déjà\n");
 
-        $this->event = new \App\Event(['name' => 'Joutes 2019', 'img' => 'joutes.jpg', 'eventState' => 0]);
+        $this->event = new \App\Models\Event(['name' => 'Joutes 2019', 'img' => 'joutes.jpg', 'eventState' => 0]);
         $this->event->save();
 
         // make room
@@ -49,17 +48,17 @@ class Joutes2020Seeder extends Seeder
     // Common stuff
     private function basics()
     {
-        (new \App\GameType(['game_type_description' => 'Modalités de jeu']))->save();
+        (new \App\Models\GameType(['game_type_description' => 'Modalités de jeu']))->save();
 
-        (new \App\PoolMode([
+        (new \App\Models\PoolMode([
             'mode_description' => 'Matches simples',
             'planningAlgorithm' => '1',
         ]))->save();
-        (new \App\PoolMode([
+        (new \App\Models\PoolMode([
             'mode_description' => 'Aller-retour',
             'planningAlgorithm' => '2',
         ]))->save();
-        (new \App\PoolMode([
+        (new \App\Models\PoolMode([
             'mode_description' => 'Elimination directe',
             'planningAlgorithm' => '3',
         ]))->save();
@@ -70,15 +69,15 @@ class Joutes2020Seeder extends Seeder
     {
         echo "Création du tournoi de beach...";
         // Beach Volley
-        $sport = new \App\Sport(['name' => 'Beach Volley', 'description' => 'Le 4-4 de la mort', 'min_participant' => 12, 'max_participant' => 12]);
+        $sport = new \App\Models\Sport(['name' => 'Beach Volley', 'description' => 'Le 4-4 de la mort', 'min_participant' => 12, 'max_participant' => 12]);
         $sport->save();
-        $court = new \App\Court(['name' => 'Lac', 'acronym' => 'LAC']);
+        $court = new \App\Models\Court(['name' => 'Lac', 'acronym' => 'LAC']);
         $court->sport()->associate($sport);
         $court->save();
-        $court = new \App\Court(['name' => 'Montagne', 'acronym' => 'MTN']);
+        $court = new \App\Models\Court(['name' => 'Montagne', 'acronym' => 'MTN']);
         $court->sport()->associate($sport);
         $court->save();
-        $tournament = new \App\Tournament(['name' => 'Beach Volley', 'start_date' => Carbon::Parse('2019-07-02 08:00'), 'end_date' => Carbon::Parse('2019-07-02 16:00'), 'max_teams' => 12, 'img' => 'beachvolley.jpg']);
+        $tournament = new \App\Models\Tournament(['name' => 'Beach Volley', 'start_date' => Carbon::Parse('2019-07-02 08:00'), 'end_date' => Carbon::Parse('2019-07-02 16:00'), 'max_teams' => 12, 'img' => 'beachvolley.jpg']);
         echo ("-");
         $tournament->sport()->associate($sport);
         echo ("-");
@@ -89,14 +88,14 @@ class Joutes2020Seeder extends Seeder
 
         echo "Création du tournoi de badminton...";
         // Bad
-        $sport = new \App\Sport(['name' => 'Badminton', 'description' => 'En double mixte (ou pas)', 'min_participant' => 12, 'max_participant' => 16]);
+        $sport = new \App\Models\Sport(['name' => 'Badminton', 'description' => 'En double mixte (ou pas)', 'min_participant' => 12, 'max_participant' => 16]);
         $sport->save();
         for ($i = 1; $i <= 6; $i++) {
-            $court = new \App\Court(['name' => 'Court ' . $i, 'acronym' => 'CT' . $i]);
+            $court = new \App\Models\Court(['name' => 'Court ' . $i, 'acronym' => 'CT' . $i]);
             $court->sport()->associate($sport);
             $court->save();
         }
-        $tournament = new \App\Tournament(['name' => 'Badminton', 'start_date' => Carbon::Parse('2019-07-02 13:30'), 'end_date' => Carbon::Parse('2019-07-02 16:00'), 'max_teams' => 16, 'img' => 'badminton.jpg']);
+        $tournament = new \App\Models\Tournament(['name' => 'Badminton', 'start_date' => Carbon::Parse('2019-07-02 13:30'), 'end_date' => Carbon::Parse('2019-07-02 16:00'), 'max_teams' => 16, 'img' => 'badminton.jpg']);
         $tournament->sport()->associate($sport);
         $tournament->event()->associate($this->event);
         $tournament->save();
@@ -104,14 +103,14 @@ class Joutes2020Seeder extends Seeder
 
         echo "Création du tournoi de basket...";
         // Basket
-        $sport = new \App\Sport(['name' => 'Basket', 'description' => 'Equipes de 5 (+1 remplaçant)', 'min_participant' => 10, 'max_participant' => 10]);
+        $sport = new \App\Models\Sport(['name' => 'Basket', 'description' => 'Equipes de 5 (+1 remplaçant)', 'min_participant' => 10, 'max_participant' => 10]);
         $sport->save();
         for ($i = 1; $i <= 2; $i++) {
-            $court = new \App\Court(['name' => 'Court ' . $i, 'acronym' => 'CT' . $i]);
+            $court = new \App\Models\Court(['name' => 'Court ' . $i, 'acronym' => 'CT' . $i]);
             $court->sport()->associate($sport);
             $court->save();
         }
-        $tournament = new \App\Tournament(['name' => 'Basket', 'start_date' => Carbon::Parse('2019-07-02 08:00'), 'end_date' => Carbon::Parse('2019-07-02 12:00'), 'max_teams' => 10, 'img' => 'basket.jpg']);
+        $tournament = new \App\Models\Tournament(['name' => 'Basket', 'start_date' => Carbon::Parse('2019-07-02 08:00'), 'end_date' => Carbon::Parse('2019-07-02 12:00'), 'max_teams' => 10, 'img' => 'basket.jpg']);
         $tournament->sport()->associate($sport);
         $tournament->event()->associate($this->event);
         $tournament->save();
@@ -119,14 +118,14 @@ class Joutes2020Seeder extends Seeder
 
         echo "Création du tournoi de unihockey...";
         // Unihockey
-        $sport = new \App\Sport(['name' => 'Unihockey', 'description' => 'Equipes de 4 (+1 remplaçant)', 'min_participant' => 10, 'max_participant' => 10]);
+        $sport = new \App\Models\Sport(['name' => 'Unihockey', 'description' => 'Equipes de 4 (+1 remplaçant)', 'min_participant' => 10, 'max_participant' => 10]);
         $sport->save();
         for ($i = 1; $i <= 2; $i++) {
-            $court = new \App\Court(['name' => 'Court ' . $i, 'acronym' => 'CT' . $i]);
+            $court = new \App\Models\Court(['name' => 'Court ' . $i, 'acronym' => 'CT' . $i]);
             $court->sport()->associate($sport);
             $court->save();
         }
-        $tournament = new \App\Tournament(['name' => 'Unihockey', 'start_date' => Carbon::Parse('2019-07-02 08:00'), 'end_date' => Carbon::Parse('2019-07-02 12:00'), 'max_teams' => 10, 'img' => 'unihockey.jpg']);
+        $tournament = new \App\Models\Tournament(['name' => 'Unihockey', 'start_date' => Carbon::Parse('2019-07-02 08:00'), 'end_date' => Carbon::Parse('2019-07-02 12:00'), 'max_teams' => 10, 'img' => 'unihockey.jpg']);
         $tournament->sport()->associate($sport);
         $tournament->event()->associate($this->event);
         $tournament->save();
@@ -134,14 +133,14 @@ class Joutes2020Seeder extends Seeder
 
         echo "Création du tournoi de pétanque...";
         // Pétanque
-        $sport = new \App\Sport(['name' => 'Pétanque', 'description' => 'Doublettes', 'min_participant' => 16, 'max_participant' => 16]);
+        $sport = new \App\Models\Sport(['name' => 'Pétanque', 'description' => 'Doublettes', 'min_participant' => 16, 'max_participant' => 16]);
         $sport->save();
         for ($i = 1; $i <= 4; $i++) {
-            $court = new \App\Court(['name' => 'Piste ' . $i, 'acronym' => 'PT' . $i]);
+            $court = new \App\Models\Court(['name' => 'Piste ' . $i, 'acronym' => 'PT' . $i]);
             $court->sport()->associate($sport);
             $court->save();
         }
-        $tournament = new \App\Tournament(['name' => 'Pétanque', 'start_date' => Carbon::Parse('2019-07-02 08:00'), 'end_date' => Carbon::Parse('2019-07-02 16:00'), 'max_teams' => 16, 'img' => 'petanque.jpg']);
+        $tournament = new \App\Models\Tournament(['name' => 'Pétanque', 'start_date' => Carbon::Parse('2019-07-02 08:00'), 'end_date' => Carbon::Parse('2019-07-02 16:00'), 'max_teams' => 16, 'img' => 'petanque.jpg']);
         $tournament->sport()->associate($sport);
         $tournament->event()->associate($this->event);
         $tournament->save();
@@ -149,14 +148,14 @@ class Joutes2020Seeder extends Seeder
 
         echo "Création du tournoi de foot...";
         // Foot
-        $sport = new \App\Sport(['name' => 'Foot', 'description' => 'A huit', 'min_participant' => 8, 'max_participant' => 8]);
+        $sport = new \App\Models\Sport(['name' => 'Foot', 'description' => 'A huit', 'min_participant' => 8, 'max_participant' => 8]);
         $sport->save();
         for ($i = 'A'; $i <= 'B'; $i++) {
-            $court = new \App\Court(['name' => 'Terrain ' . $i, 'acronym' => 'T' . $i]);
+            $court = new \App\Models\Court(['name' => 'Terrain ' . $i, 'acronym' => 'T' . $i]);
             $court->sport()->associate($sport);
             $court->save();
         }
-        $tournament = new \App\Tournament(['name' => 'Foot', 'start_date' => Carbon::Parse('2019-07-02 13:30'), 'end_date' => Carbon::Parse('2019-07-02 16:00'), 'max_teams' => 8, 'img' => 'football.jpg']);
+        $tournament = new \App\Models\Tournament(['name' => 'Foot', 'start_date' => Carbon::Parse('2019-07-02 13:30'), 'end_date' => Carbon::Parse('2019-07-02 16:00'), 'max_teams' => 8, 'img' => 'football.jpg']);
         $tournament->sport()->associate($sport);
         $tournament->event()->associate($this->event);
         $tournament->save();
@@ -164,12 +163,12 @@ class Joutes2020Seeder extends Seeder
 
         echo "Création de la marche...";
         // Marche
-        $sport = new \App\Sport(['name' => 'Marche', 'description' => 'Par là autour', 'min_participant' => 1, 'max_participant' => 100]);
+        $sport = new \App\Models\Sport(['name' => 'Marche', 'description' => 'Par là autour', 'min_participant' => 1, 'max_participant' => 100]);
         $sport->save();
-        $court = new \App\Court(['name' => 'Sainte-Croix et alentours', 'acronym' => 'Stex']);
+        $court = new \App\Models\Court(['name' => 'Sainte-Croix et alentours', 'acronym' => 'Stex']);
         $court->sport()->associate($sport);
         $court->save();
-        $tournament = new \App\Tournament(['name' => 'Marche', 'start_date' => Carbon::Parse('2019-07-02 08:00'), 'end_date' => Carbon::Parse('2019-07-02 16:00'), 'max_teams' => 100, 'img' => 'rando.jpg']);
+        $tournament = new \App\Models\Tournament(['name' => 'Marche', 'start_date' => Carbon::Parse('2019-07-02 08:00'), 'end_date' => Carbon::Parse('2019-07-02 16:00'), 'max_teams' => 100, 'img' => 'rando.jpg']);
         $tournament->sport()->associate($sport);
         $tournament->event()->associate($this->event);
         $tournament->save();
@@ -259,7 +258,7 @@ class Joutes2020Seeder extends Seeder
     private function Badminton()
     {
         echo "Badminton\n";
-        $bv = \App\Tournament::where('name', 'like', '%Badmin%')->first();
+        $bv = \App\Models\Tournament::where('name', 'like', '%Badmin%')->first();
         if (!$bv) {
             echo "Le tournoi de badminton n'existe pas\n";
             return;
@@ -267,9 +266,9 @@ class Joutes2020Seeder extends Seeder
         $tournamentid = $bv->id;
         $sportid = $bv->sport_id;
 
-        $teams = \App\Team::where('tournament_id', '=', $tournamentid)->get();
+        $teams = \App\Models\Team::where('tournament_id', '=', $tournamentid)->get();
 
-        $poolModeId = \App\PoolMode::where('mode_description', '=', 'Matches simples')->get()->first()->id;
+        $poolModeId = \App\Models\PoolMode::where('mode_description', '=', 'Matches simples')->get()->first()->id;
 
         $gameTypeId = GameType::where('game_type_description', '=', 'Modalités de jeu')->get()->first()->id;
 
@@ -277,7 +276,7 @@ class Joutes2020Seeder extends Seeder
         //================================================================================================================
         echo "Championnat\n";
 
-        $pool = new \App\Pool([
+        $pool = new \App\Models\Pool([
             'tournament_id' => $tournamentid,
             'start_time' => '09:30',
             'end_time' => '16:00',
@@ -293,14 +292,14 @@ class Joutes2020Seeder extends Seeder
 
 
         foreach ($teams as $team) {
-            (new \App\Contender([
+            (new \App\Models\Contender([
                 'pool_id' => $firstpoolStage1,
                 'team_id' => $team->id
             ]))->save();
         }
-        $firstcontender = \App\Contender::where('pool_id', '=', $firstpoolStage1)->first()->id;
+        $firstcontender = \App\Models\Contender::where('pool_id', '=', $firstpoolStage1)->first()->id;
 
-        $firstcourt = \App\Court::where('sport_id', '=', $sportid)->first()->id;
+        $firstcourt = \App\Models\Court::where('sport_id', '=', $sportid)->first()->id;
 
         $teams = array(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12); // Offsets from database id of first contender
 
@@ -326,7 +325,7 @@ class Joutes2020Seeder extends Seeder
             $team2Index = $evenNumberOfTeams ? $nbTeams - 2 : $nbTeams - 1;
             // "draw the horizontal lines in the polygon", leaving the first team out
             while ($team1Index < $team2Index) {
-                (new \App\Game(['date' => '2018-07-03', 'start_time' => "$poolHour:$poolMinute", 'contender1_id' => $firstcontender + $teams[$team1Index], 'contender2_id' => $firstcontender + $teams[$team2Index], 'court_id' => $firstcourt]))->save();
+                (new \App\Models\Game(['date' => '2018-07-03', 'start_time' => "$poolHour:$poolMinute", 'contender1_id' => $firstcontender + $teams[$team1Index], 'contender2_id' => $firstcontender + $teams[$team2Index], 'court_id' => $firstcourt]))->save();
                 $team1Index++;
                 $team2Index--;
             }
@@ -352,7 +351,7 @@ class Joutes2020Seeder extends Seeder
     private function UniHockey()
     {
         echo "Unihockey\n";
-        $bv = \App\Tournament::where('name', 'like', '%hockey%')->first();
+        $bv = \App\Models\Tournament::where('name', 'like', '%hockey%')->first();
         if (!$bv) {
             echo "Le tournoi de unihockey n'existe pas\n";
             return;
@@ -360,9 +359,9 @@ class Joutes2020Seeder extends Seeder
         $tournamentid = $bv->id;
         $sportid = $bv->sport_id;
 
-        $teams = \App\Team::where('tournament_id', '=', $tournamentid)->get();
+        $teams = \App\Models\Team::where('tournament_id', '=', $tournamentid)->get();
 
-        $poolModeId = \App\PoolMode::where('mode_description', '=', 'Matches simples')->get()->first()->id;
+        $poolModeId = \App\Models\PoolMode::where('mode_description', '=', 'Matches simples')->get()->first()->id;
 
         $gameTypeId = GameType::where('game_type_description', '=', 'Modalités de jeu')->get()->first()->id;
 
@@ -370,7 +369,7 @@ class Joutes2020Seeder extends Seeder
         //================================================================================================================
         echo "Championnat\n";
 
-        $pool = new \App\Pool([
+        $pool = new \App\Models\Pool([
             'tournament_id' => $tournamentid,
             'start_time' => '09:30',
             'end_time' => '16:00',
@@ -386,44 +385,44 @@ class Joutes2020Seeder extends Seeder
 
 
         foreach ($teams as $team) {
-            (new \App\Contender([
+            (new \App\Models\Contender([
                 'pool_id' => $firstpoolStage1,
                 'team_id' => $team->id
             ]))->save();
         }
-        $firstcontender = \App\Contender::where('pool_id', '=', $firstpoolStage1)->first()->id;
+        $firstcontender = \App\Models\Contender::where('pool_id', '=', $firstpoolStage1)->first()->id;
 
-        $firstcourt = \App\Court::where('sport_id', '=', $sportid)->first()->id;
+        $firstcourt = \App\Models\Court::where('sport_id', '=', $sportid)->first()->id;
 
         // Thank you https://nrich.maths.org/1443
         // Games of pool A
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:51', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:51', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:12', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:12', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:33', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:33', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:54', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:54', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:15', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:15', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:00', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:00', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:21', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:21', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:42', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:42', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:03', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:03', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:24', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:51', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:51', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:12', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:12', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:33', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:33', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:54', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:54', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:15', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:15', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:00', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:00', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:21', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:21', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:42', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:42', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:03', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:03', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:24', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
     }
 
     private function Basket()
     {
         echo "Basket\n";
-        $bv = \App\Tournament::where('name', 'like', '%Basket%')->first();
+        $bv = \App\Models\Tournament::where('name', 'like', '%Basket%')->first();
         if (!$bv) {
             echo "Le tournoi de basket n'existe pas\n";
             return;
@@ -431,9 +430,9 @@ class Joutes2020Seeder extends Seeder
         $tournamentid = $bv->id;
         $sportid = $bv->sport_id;
 
-        $teams = \App\Team::where('tournament_id', '=', $tournamentid)->get();
+        $teams = \App\Models\Team::where('tournament_id', '=', $tournamentid)->get();
 
-        $poolModeId = \App\PoolMode::where('mode_description', '=', 'Matches simples')->get()->first()->id;
+        $poolModeId = \App\Models\PoolMode::where('mode_description', '=', 'Matches simples')->get()->first()->id;
 
         $gameTypeId = GameType::where('game_type_description', '=', 'Modalités de jeu')->get()->first()->id;
 
@@ -441,7 +440,7 @@ class Joutes2020Seeder extends Seeder
         //================================================================================================================
         echo "Championnat\n";
 
-        $pool = new \App\Pool([
+        $pool = new \App\Models\Pool([
             'tournament_id' => $tournamentid,
             'start_time' => '09:30',
             'end_time' => '16:00',
@@ -457,57 +456,57 @@ class Joutes2020Seeder extends Seeder
 
 
         foreach ($teams as $team) {
-            (new \App\Contender([
+            (new \App\Models\Contender([
                 'pool_id' => $firstpoolStage1,
                 'team_id' => $team->id
             ]))->save();
         }
-        $firstcontender = \App\Contender::where('pool_id', '=', $firstpoolStage1)->first()->id;
+        $firstcontender = \App\Models\Contender::where('pool_id', '=', $firstpoolStage1)->first()->id;
 
-        $firstcourt = \App\Court::where('sport_id', '=', $sportid)->first()->id;
+        $firstcourt = \App\Models\Court::where('sport_id', '=', $sportid)->first()->id;
 
         // Thank you https://nrich.maths.org/1443
         // Games of pool A
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:47', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:04', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:21', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:47', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:04', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:21', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt]))->save();
 
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:38', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:55', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:12', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:29', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:38', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:55', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:12', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:29', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt]))->save();
 
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:46', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '12:03', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '12:20', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '12:37', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:46', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '12:03', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '12:20', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '12:37', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt]))->save();
 
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '12:54', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:11', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:28', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:45', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '12:54', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:11', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:28', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:45', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
 
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:02', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:19', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:36', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:53', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:02', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:19', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:36', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:53', 'contender1_id' => $firstcontender + 5, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
 
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:10', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:27', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:44', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '16:01', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:10', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 0, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:27', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:44', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '16:01', 'contender1_id' => $firstcontender + 4, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
 
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '16:18', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '16:35', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '16:52', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '17:09', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '16:18', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '16:35', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '16:52', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '17:09', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
     }
 
     private function BeachVolley()
     {
         echo "Beach Volley\n";
-        $bv = \App\Tournament::where('name', 'like', '%Beach%')->first();
+        $bv = \App\Models\Tournament::where('name', 'like', '%Beach%')->first();
         if (!$bv) {
             echo "Le tournoi de beach n'existe pas\n";
             return;
@@ -515,9 +514,9 @@ class Joutes2020Seeder extends Seeder
         $tournamentid = $bv->id;
         $sportid = $bv->sport_id;
 
-        $teams = \App\Team::where('tournament_id', '=', $tournamentid)->get();
+        $teams = \App\Models\Team::where('tournament_id', '=', $tournamentid)->get();
 
-        $poolModeId = \App\PoolMode::where('mode_description', '=', 'Matches simples')->get()->first()->id;
+        $poolModeId = \App\Models\PoolMode::where('mode_description', '=', 'Matches simples')->get()->first()->id;
 
         $gameTypeId = GameType::where('game_type_description', '=', 'Modalités de jeu')->get()->first()->id;
 
@@ -525,7 +524,7 @@ class Joutes2020Seeder extends Seeder
         //================================================================================================================
         echo "Stage 1 = 2 poules de 6 équipes\n";
 
-        $pool = new \App\Pool([
+        $pool = new \App\Models\Pool([
             'tournament_id' => $tournamentid,
             'start_time' => '09:30',
             'end_time' => '11:45',
@@ -539,7 +538,7 @@ class Joutes2020Seeder extends Seeder
         $pool->save();
         $firstpoolStage1 = $pool->id; // we'll need that to put teams into pools
 
-        (new \App\Pool([
+        (new \App\Models\Pool([
             'tournament_id' => $tournamentid,
             'start_time' => '09:30',
             'end_time' => '11:45',
@@ -553,53 +552,53 @@ class Joutes2020Seeder extends Seeder
 
         $nbTeams = 0;
         foreach ($teams as $team) {
-            (new \App\Contender([
+            (new \App\Models\Contender([
                 'pool_id' => $firstpoolStage1 + intdiv($nbTeams, 6),
                 'team_id' => $team->id
             ]))->save();
             $nbTeams++;
         }
-        $firstcontender = \App\Contender::where('pool_id', '=', $firstpoolStage1)->first()->id;
-        $firstcourt = \App\Court::where('sport_id', '=', $sportid)->first()->id;
+        $firstcontender = \App\Models\Contender::where('pool_id', '=', $firstpoolStage1)->first()->id;
+        $firstcourt = \App\Models\Court::where('sport_id', '=', $sportid)->first()->id;
 
         // Games of pool A
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:40', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:50', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:00', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:10', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:20', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:40', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:50', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:00', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:10', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:20', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:40', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:50', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:40', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:50', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:00', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:10', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:20', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:40', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:50', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:00', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:10', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:20', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:40', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:50', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
 
         // Games of pool B
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:40', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '09:50', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:00', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 8, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:10', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:20', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:40', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '10:50', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:00', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:10', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 8, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:20', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:40', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '11:50', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:40', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '09:50', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:00', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 8, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:10', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:20', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:40', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '10:50', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:00', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:10', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 8, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:20', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:40', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '11:50', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
 
         //================================================================================================================
         echo "Stage 2 = 2 poules de 6 équipes\n";
 
-        $pool = new \App\Pool([
+        $pool = new \App\Models\Pool([
             'tournament_id' => $tournamentid,
             'start_time' => '09:30',
             'end_time' => '11:45',
@@ -613,7 +612,7 @@ class Joutes2020Seeder extends Seeder
         $pool->save();
         $firstpoolStage2 = $pool->id; // we'll need that to put teams into pools
 
-        (new \App\Pool([
+        (new \App\Models\Pool([
             'tournament_id' => $tournamentid,
             'start_time' => '09:30',
             'end_time' => '11:45',
@@ -625,56 +624,56 @@ class Joutes2020Seeder extends Seeder
             'poolState' => 1
         ]))->save();
 
-        (new \App\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 1]))->save();
-        (new \App\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 2]))->save();
-        (new \App\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 3]))->save();
-        (new \App\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 1]))->save();
-        (new \App\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 2]))->save();
-        (new \App\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 3]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 1]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 2]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 3]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 1]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 2]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 3]))->save();
 
-        (new \App\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 4]))->save();
-        (new \App\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 5]))->save();
-        (new \App\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 6]))->save();
-        (new \App\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 4]))->save();
-        (new \App\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 5]))->save();
-        (new \App\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 6]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 4]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 5]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1, 'rank_in_pool' => 6]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 4]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 5]))->save();
+        (new \App\Models\Contender(['pool_id' => $firstpoolStage2 + 1, 'pool_from_id' => $firstpoolStage1 + 1, 'rank_in_pool' => 6]))->save();
 
-        $firstcontender = \App\Contender::where('pool_id', '=', $firstpoolStage2)->first()->id;
-        $firstcourt = \App\Court::where('sport_id', '=', $sportid)->first()->id;
+        $firstcontender = \App\Models\Contender::where('pool_id', '=', $firstpoolStage2)->first()->id;
+        $firstcourt = \App\Models\Court::where('sport_id', '=', $sportid)->first()->id;
 
         // Games of pool Winner
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:00', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:10', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:20', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:40', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:50', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:00', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:10', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:20', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:40', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:50', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:00', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:10', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:20', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:00', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 1, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:10', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:20', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:40', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:50', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:00', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:10', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:20', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:30', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:40', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 2, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:50', 'contender1_id' => $firstcontender + 3, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:00', 'contender1_id' => $firstcontender + 0, 'contender2_id' => $firstcontender + 5, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:10', 'contender1_id' => $firstcontender + 1, 'contender2_id' => $firstcontender + 3, 'court_id' => $firstcourt]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:20', 'contender1_id' => $firstcontender + 2, 'contender2_id' => $firstcontender + 4, 'court_id' => $firstcourt]))->save();
 
         // Games of pool Losers
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:00', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:10', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:20', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 8, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:40', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '13:50', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:00', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:10', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:20', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:40', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 8, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '14:50', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:00', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:10', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
-        (new \App\Game(['date' => '2017-06-27', 'start_time' => '15:20', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:00', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 7, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:10', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:20', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 8, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:40', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '13:50', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:00', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:10', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 6, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:20', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:30', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:40', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 8, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '14:50', 'contender1_id' => $firstcontender + 9, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:00', 'contender1_id' => $firstcontender + 6, 'contender2_id' => $firstcontender + 11, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:10', 'contender1_id' => $firstcontender + 7, 'contender2_id' => $firstcontender + 9, 'court_id' => $firstcourt + 1]))->save();
+        (new \App\Models\Game(['date' => '2017-06-27', 'start_time' => '15:20', 'contender1_id' => $firstcontender + 8, 'contender2_id' => $firstcontender + 10, 'court_id' => $firstcourt + 1]))->save();
     }
 
 
@@ -682,15 +681,15 @@ class Joutes2020Seeder extends Seeder
     {
         echo "Création des roles utilisateurs\n";
         // Admin
-        $role = new \App\Role(['name' => 'Administrateur', 'slug' => 'ADMIN',]);
+        $role = new \App\Models\Role(['name' => 'Administrateur', 'slug' => 'ADMIN',]);
         $role->save();
 
         // Gestionnaire
-        $role = new \App\Role(['name' => 'Gestionnaire', 'slug' => 'GEST',]);
+        $role = new \App\Models\Role(['name' => 'Gestionnaire', 'slug' => 'GEST',]);
         $role->save();
 
         // Élève
-        $role = new \App\Role(['name' => 'Student', 'slug' => 'STUD',]);
+        $role = new \App\Models\Role(['name' => 'Student', 'slug' => 'STUD',]);
         $role->save();
 
         echo "OK\n";
@@ -699,7 +698,7 @@ class Joutes2020Seeder extends Seeder
     private function users()
     {
         echo "Création d'un utilisateur ...";
-        $this->user = new \App\User([
+        $this->user = new \App\Models\User([
             'username' => 'AdminTest',
             'email' => 'Admin@cpnv.ch',
             'first_name' => 'Admin',
