@@ -17,7 +17,7 @@ class UserTableSeeder extends Seeder
     public function run()
     {
         $faker = Faker::create();
-        foreach (range(1, 200) as $index) {
+        foreach (range(1, 10) as $index) {
             DB::table('users')->insert([
                 'username' => $faker->userName,
                 'password' => bcrypt('secret'),
