@@ -20,6 +20,11 @@
 
                 <h2>Matches et Résultats</h2>
                 <h4>État: {{ \App\Enums\PoolState::poolStateName($pool->poolState) }}</h4>
+                @if ($pool->isEditable() && Helper::eventStateName($pool->tournament->event->eventState) == 'En cours')
+                    <button type="submit" class="btn btn-main" data-toggle="modal" data-target="#stagePoolModal">Passer à l'étape suivante : {{ \App\Enums\PoolState::poolStateName($pool->poolState + 1) }}</button>
+                @else
+                    <h5>En attente de l'activation de l'évènement</h5>
+                @endif
                 <h4>Date : {{ $tournament->start_date->format('d.m.Y') }}</h4>
             </div>
         </div>
@@ -38,7 +43,13 @@
                                     @if (!isset($game->score_contender1) && !isset($game->score_contender2))
                                         <tr>
                                             <td class="separator sepTime ">
-                                                {{ Carbon\Carbon::parse($game->start_time)->format('H:i') }}</td>
+                                            @if ($pool->isEditable())
+                                                <td><input type="time" name="game[{{ $game->id }}][editedTime]"
+                                                           value="{{ Carbon\Carbon::parse($game->start_time)->format('H:i') }}">
+                                                </td>
+                                            @else
+                                                <td>{{ Carbon\Carbon::parse($game->start_time)->format('H:i') }}</td>
+                                            @endif
                                             <td class="contender1 ">
                                                 @if (empty($game->contender1->team))
                                                     {{ $game->contender1->rank_in_pool . ($game->contender1->rank_in_pool == 1 ? 'er ' : 'ème ') . 'de ' . $game->contender1->fromPool->poolName }}
@@ -205,7 +216,7 @@
                 <div class="modal-footer">
                     <form action="{{ route('tournaments.pools.update', [$tournament, $pool]) }}" method="POST">
                         @csrf
-                        <input hidden type="number" value="1" name="poolState">
+                        <input hidden type="number" value="3" name="poolState">
                         <button type="button" class="btn btn-secondary" data-dismiss="modal">Annuler</button>
                         <input type="hidden" name="_method" value="PATCH">
                         <button type="submit" name="changeStatePool" class="btn btn-success">Ok !</button>
@@ -214,4 +225,16 @@
             </div>
         </div>
     </div>
+    @section('customScript')
+        <script>
+            $("input:regex(name, game\\[\\d*\\]\\[editedTime\\])").each((index, element)=>{
+                let items = $("input:regex(name, game\\[\\d*\\]\\[editedTime\\])")
+                $(element).change((event) => {
+                    for (let itemIndex = index + 1; itemIndex < items.length; itemIndex++){
+                        items[itemIndex].value = Time($(element).val()).addMinutes((itemIndex - index) * 10).toString();
+                    }
+                })
+            })
+        </script>
+    @stop
 @stop
