@@ -5,7 +5,7 @@
         <div class="row">
             <div class="col-md-2">
                 <a href="{{ route('tournaments.show', $tournament) }}"><i
-                        class="fa fa-4x fa-arrow-circle-left return fa-return growIcon" aria-hidden="true"></i></a>
+                            class="fa fa-4x fa-arrow-circle-left return fa-return growIcon" aria-hidden="true"></i></a>
                 <button type="submit" class="btn btn-success" data-toggle="modal" data-target="#updateModal">
                     <i class="fa fa-edit fa-1x" aria-hidden="true"></i>
                 </button>
@@ -20,8 +20,7 @@
                 <h2>Matches et Résultats</h2>
                 <h4>État: {{ \App\Enums\PoolState::poolStateName($pool->poolState) }}</h4>
                 @if ($pool->isEditable() && Helper::eventStateName($pool->tournament->event->eventState) == 'En cours')
-                    <button type="submit" class="btn btn-main" data-toggle="modal" data-target="#stagePoolModal">Passer à
-                        l'étape suivante : {{ \App\Enums\PoolState::poolStateName($pool->poolState + 1) }}</button>
+                    <button type="submit" class="btn btn-main" data-toggle="modal" data-target="#stagePoolModal">Passer à l'étape suivante : {{ \App\Enums\PoolState::poolStateName($pool->poolState + 1) }}</button>
                 @else
                     <h5>En attente de l'activation de l'évènement</h5>
                 @endif
@@ -34,41 +33,38 @@
                 <input type="hidden" name="_method" value="PUT">
                 <table class="table" id="data_table">
                     <tbody class="text">
-                        @foreach ($games as $game)
-                            <tr>
-                                @if (empty($game->contender1->team) || empty($game->contender2->team))
-                                    <td class="separator sepTime " id="time{{ $game->id }}">
-                                        {{ Carbon\Carbon::parse($game->start_time)->format('H:i') }}</td>
-                                    <td class="contender1">
-                                        {{ $game->contender1->rank_in_pool .($game->contender1->rank_in_pool == 1 ? 'er ' : 'ème ') .'de ' .$game->contender1->fromPool->poolName }}
-                                    </td>
-                                    <td class="separator"> - </td>
+                    @foreach ($games as $game)
+                        <tr>
+                            @if (empty($game->contender1->team) || empty($game->contender2->team))
+                                <td class="separator sepTime " id="time{{ $game->id }}">
+                                    {{ Carbon\Carbon::parse($game->start_time)->format('H:i') }}</td>
+                                <td class="contender1">
+                                    {{ $game->contender1->rank_in_pool .($game->contender1->rank_in_pool == 1 ? 'er ' : 'ème ') .'de ' .$game->contender1->fromPool->poolName }}
+                                </td>
+                                <td class="separator"> -</td>
 
-                                    <td class="contender2">
-                                        {{ $game->contender2->rank_in_pool .($game->contender2->rank_in_pool == 1 ? 'er ' : 'ème ') .'de ' .$game->contender2->fromPool->poolName }}
-                                    </td>
-                                    <td class="score2" id="court{{ $game->id }}">{{ $game->court->name }}
-                                    </td>
-                                @else
-                                    <td class="separator sepTime " id="time{{ $game->id }}">
-                                        {{ Carbon\Carbon::parse($game->start_time)->format('H:i') }}</td>
-                                    <td class="contender1">{{ $game->contender1->team->name }}</td>
-                                    <td class="separator"> - </td>
+                                <td class="contender2">
+                                    {{ $game->contender2->rank_in_pool .($game->contender2->rank_in_pool == 1 ? 'er ' : 'ème ') .'de ' .$game->contender2->fromPool->poolName }}
+                                </td>
+                                <td class="score2" id="court{{ $game->id }}">{{ $game->court->name }}
+                                </td>
+                            @else
+                                <td class="separator sepTime " id="time{{ $game->id }}">
+                                    {{ Carbon\Carbon::parse($game->start_time)->format('H:i') }}</td>
+                                <td class="contender1">{{ $game->contender1->team->name }}</td>
+                                <td class="separator"> -</td>
 
-                                    <td class="contender2">{{ $game->contender2->team->name }}</td>
-                                    <td class="score2" id="court{{ $game->id }}">{{ $game->court->name }}
-                                    </td>
-                                @endif
-                            </tr>
-                        @endforeach
+                                <td class="contender2">{{ $game->contender2->team->name }}</td>
+                                <td class="score2" id="court{{ $game->id }}">{{ $game->court->name }}
+                                </td>
+                            @endif
+                        </tr>
+                    @endforeach
                     </tbody>
                 </table>
 
 
-
-
-
-                @if (\App\Contender::isAllEmpty($contenders))
+                @if (\App\Models\Contender::isAllEmpty($contenders))
                     <h2>Liste des participants</h2>
                     <table class="table">
                         @foreach ($contenders as $contender)
@@ -93,7 +89,7 @@
 
     <!-- Modal Deletion -->
     <div class="modal fade" id="deleteModal" tabindex="-1" role="dialog" aria-labelledby="deleteModalLabel"
-        aria-hidden="true">
+         aria-hidden="true">
         <div class="modal-dialog" role="document">
             <div class="modal-content">
 
@@ -119,7 +115,7 @@
     </div>
     <!-- Modal Update -->
     <div class="modal fade" id="updateModal" tabindex="-1" role="dialog" aria-labelledby="updateModalLabel"
-        aria-hidden="true">
+         aria-hidden="true">
         <div class="modal-dialog" role="document">
             <div class="modal-content">
 
@@ -146,7 +142,7 @@
     </div>
     <!-- Modal StagePool -->
     <div class="modal fade" id="stagePoolModal" tabindex="-1" role="dialog" aria-labelledby="stagePoolModalLabel"
-        aria-hidden="true">
+         aria-hidden="true">
         <div class="modal-dialog" role="document">
             <div class="modal-content">
 
